album rating format %.1f/5

+++ app/src/main/java/phramusca/com/jamuzremote/ActivityMain.java
@@ -2124,7 +2124,7 @@ public class ActivityMain extends AppCompatActivity {
                         .concat(displayedTrack.toString())
                         .concat("</html>"))));
                 ratingBar.setEnabled(false);
-                ratingBar.setRating(displayedTrack.getRating());
+                ratingBar.setRating((float)displayedTrack.getRating());
                 ratingBar.setEnabled(true);
                 setupSpinnerGenre(RepoGenres.get(), displayedTrack.getGenre());
